Improvements by David Bielejeski, Niko Pueringer, Maki99
